Refresh S3 credentials in case Log Processing role is deleted (#2642)
